+++ src/nodes/arrays.js
@@ -1,4 +1,4 @@
-const { concat, group, ifBreak, indent, join, line, softline } = require("prettier").doc.builders;
+const { concat, group, ifBreak, indent, join, line, literalline, softline } = require("prettier").doc.builders;
 
 const isStringArray = args => args.body.every(arg => (
   arg.type === "string_literal"
@@ -36,6 +36,25 @@ const printSpecialArray = parts => group(concat([
   concat([softline, "]"])
 ]));
 
+// Extract out the actual elements, taking into account nesting with
+// `args_add_star` nodes. The only nodes that get passed into this function are
+// `args` or `args_add_star`.
+const getElements = (node, elementPath) => {
+  if (node.type === "args") {
+    return node.body.map((element, index) => ({
+      element,
+      elementPath: elementPath.concat(["body", index])
+    }));
+  }
+
+  return getElements(node.body[0], elementPath.concat(["body", 0])).concat(
+    node.body.slice(1).map((element, index) => ({
+      element,
+      elementPath: elementPath.concat(["body", index + 1])
+    }))
+  );
+};
+
 module.exports = {
   aref: (path, opts, print) => {
     if (!path.getValue().body[1]) {
@@ -60,19 +79,56 @@ module.exports = {
       return printSpecialArray(["%i"].concat(getSpecialArrayParts(path, print, args)));
     }
 
-    if (["args", "args_add_star"].includes(args.type)) {
-      return group(concat([
-        "[",
-        indent(concat([
-          softline,
-          join(concat([",", line]), path.call(print, "body", 0)),
-          addTrailingCommas ? ifBreak(",", "") : ""
-        ])),
-        concat([softline, "]"])
-      ]));
+    if (!["args", "args_add_star"].includes(args.type)) {
+      return printSpecialArray(path.call(print, "body", 0));
     }
 
-    return printSpecialArray(path.call(print, "body", 0));
+    const normalDocs = [];
+
+    const elementDocs = path.call(print, "body", 0);
+    const elements = getElements(path.getValue().body[0], ["body", 0]);
+
+    // We need to manually loop through the elements in the array in order to
+    // take care of heredocs printing (their commas go after the opening, as
+    // opposed to at the end).
+    elements.forEach(({ element, elementPath }, index) => {
+      const isInner = index !== elements.length - 1;
+
+      const isStraightHeredoc = element.type === "heredoc";
+      const isSquigglyHeredoc = element.type === "string_literal" && element.body[0].type === "heredoc";
+
+      if (isStraightHeredoc || isSquigglyHeredoc) {
+        const heredocNode = isStraightHeredoc ? element : element.body[0];
+        const heredocPath = [print].concat(elementPath);
+
+        if (isSquigglyHeredoc) {
+          heredocPath.push("body", 0);
+        }
+
+        normalDocs.push(
+          heredocNode.beging,
+          (isInner || addTrailingCommas) ? "," : "",
+          literalline,
+          concat(path.map.apply(path, heredocPath.concat("body"))),
+          heredocNode.ending,
+          isInner ? line : ""
+        );
+      } else {
+        normalDocs.push(elementDocs[index]);
+
+        if (isInner) {
+          normalDocs.push(concat([",", line]));
+        } else if (addTrailingCommas) {
+          normalDocs.push(ifBreak(",", ""));
+        }
+      }
+    });
+
+    return group(concat([
+      "[",
+      indent(concat([softline].concat(normalDocs))),
+      concat([softline, "]"])
+    ]));
   },
   qsymbols: makeArray("%i"),
   qwords: makeArray("%w"),