+++ src/nodes/hashes.js
@@ -1,68 +1,106 @@
-const { concat, group, ifBreak, indent, join, line } = require("prettier").doc.builders;
-const { printComments, skipAssignIndent } = require("../utils");
+const { concat, group, ifBreak, indent, join, line, literalline } = require("prettier").doc.builders;
+const { skipAssignIndent } = require("../utils");
+
+const nodeDive = (node, steps) => {
+  let current = node;
+
+  steps.forEach(step => {
+    current = current[step];
+  });
+
+  return current;
+};
+
+const makeLabel = (path, { preferHashLabels }, print, steps) => {
+  const labelNode = nodeDive(path.getValue(), steps);
+  const labelDoc = path.call.apply(path, [print].concat(steps));
+
+  switch (labelNode.type) {
+    case "@label":
+      if (preferHashLabels) {
+        return labelDoc;
+      }
+      return `:${labelDoc.slice(0, labelDoc.length - 1)} =>`;
+    case "symbol_literal":
+      if (preferHashLabels && labelNode.body.length === 1) {
+        const symbolSteps = steps.concat("body", 0, "body", 0);
+        return concat([path.call.apply(path, [print].concat(symbolSteps)), ":"]);
+      }
+      return concat([labelDoc, " =>"]);
+    case "dyna_symbol":
+      if (preferHashLabels) {
+        return concat(labelDoc.parts.slice(1).concat(":"));
+      }
+      return concat([labelDoc, " =>"]);
+    default:
+      return concat([labelDoc, " =>"]);
+  }
+};
 
 module.exports = {
-  assoc_new: (path, { preferHashLabels }, print) => {
-    const parts = [];
-    const [printedLabel, printedValue] = path.map(print, "body");
-
-    switch (path.getValue().body[0].type) {
-      case "@label":
-        if (preferHashLabels) {
-          parts.push(printedLabel);
-        } else {
-          parts.push(`:${printedLabel.slice(0, printedLabel.length - 1)} =>`);
-        }
-        break;
-      case "symbol_literal":
-        if (preferHashLabels && path.getValue().body[0].body.length === 1) {
-          const { comments, start } = path.getValue().body[0];
-          const node = concat([path.call(print, "body", 0, "body", 0, "body", 0), ":"]);
-
-          if (comments) {
-            parts.push(printComments(node, start, comments));
-          } else {
-            parts.push(node);
-          }
-        } else {
-          parts.push(concat([printedLabel, " =>"]));
-        }
-        break;
-      case "dyna_symbol":
-        if (preferHashLabels) {
-          parts.push(concat(printedLabel.parts.slice(1).concat(":")));
-        } else {
-          parts.push(concat([printedLabel, " =>"]));
-        }
-        break;
-      default:
-        parts.push(concat([printedLabel, " =>"]));
-        break;
-    }
+  assoc_new: (path, opts, print) => {
+    const valueDoc = path.call(print, "body", 1);
+    const parts = [makeLabel(path, opts, print, ["body", 0])];
 
     if (skipAssignIndent(path.getValue().body[1])) {
-      parts.push(" ", printedValue);
+      parts.push(" ", valueDoc);
     } else {
-      parts.push(indent(concat([line, printedValue])));
+      parts.push(indent(concat([line, valueDoc])));
     }
 
     return group(concat(parts));
   },
+  assoclist_from_args: (path, opts, print) => {
+    const { addTrailingCommas } = opts;
+
+    const assocNodes = path.getValue().body[0];
+    const assocDocs = [];
+
+    assocNodes.forEach((assocNode, index) => {
+      const isInner = index !== assocNodes.length - 1;
+      const valueNode = assocNode.body[1];
+
+      const isStraightHeredoc = valueNode.type === "heredoc";
+      const isSquigglyHeredoc = valueNode.type === "string_literal" && valueNode.body[0].type === "heredoc";
+
+      if (isStraightHeredoc || isSquigglyHeredoc) {
+        const heredocSteps = isStraightHeredoc ? ["body", 1] : ["body", 1, "body", 0];
+        const { beging, ending } = nodeDive(assocNode, heredocSteps);
+
+        assocDocs.push(
+          makeLabel(path, opts, print, ["body", 0, index, "body", 0]),
+          " ",
+          beging,
+          (isInner || addTrailingCommas) ? "," : "",
+          literalline,
+          concat(path.map.apply(path, [print, "body", 0, index].concat(heredocSteps).concat("body"))),
+          ending,
+          isInner ? line : ""
+        );
+      } else {
+        assocDocs.push(path.call(print, "body", 0, index));
+
+        if (isInner) {
+          assocDocs.push(concat([",", line]));
+        } else if (addTrailingCommas) {
+          assocDocs.push(ifBreak(",", ""));
+        }
+      }
+    });
+
+    return group(concat(assocDocs));
+  },
   bare_assoc_hash: (path, opts, print) => group(
     join(concat([",", line]), path.map(print, "body", 0))
   ),
-  hash: (path, { addTrailingCommas }, print) => {
+  hash: (path, opts, print) => {
     if (path.getValue().body[0] === null) {
       return "{}";
     }
 
     return group(concat([
       "{",
-      indent(concat([
-        line,
-        concat(path.map(print, "body")),
-        addTrailingCommas ? ifBreak(",", "") : ""
-      ])),
+      indent(concat([line, concat(path.map(print, "body"))])),
       concat([line, "}"])
     ]));
   }