+++ test.rb
@@ -1,0 +1,7 @@
+def foo
+  begin
+    a
+  rescue
+    b
+  end
+end